fix: resolve remaining clippy warnings
Eliminates the last 27 clippy cosmetic warnings, bringing the kr64
library to zero clippy warnings across --lib and --all-targets.

Fixes by category:
- doc_overindented_list_items (lib.rs, seccomp.rs): reduce continuation
  indent from 8/11 spaces to the 5-space canonical minimum.
- doc_lazy_continuation (lib.rs, battery.rs): replace '+' markers that
  were being parsed as nested list bullets with plain prose so the
  continuation line is not flagged.
- unnecessary_cast (binder.rs x3, seccomp.rs x1): drop the redundant
  'as i32'/'as u32' on libc::EINVAL and SECCOMP_FILTER_FLAG_TSYNC.
- io_other_error (binder.rs x2): modernize the verbose
  io::Error::new(io::ErrorKind::Other, format!(...)) pattern to the
  idiomatic io::Error::other(format!(...)) (stabilised Rust 1.74).
- needless_borrows_for_generic_args (devices.rs): drop the '&' on the
  '&[0u8; 8]' argument to fs::write.
- manual_clamp (audio.rs x2): replace .min(MAX).max(MIN) with
  .clamp(MIN, MAX) for clarity.
- non_canonical_clone_impl (sensors.rs): replace the field-by-field
  packed-struct clone with the canonical '*self' (sound because
  taking '&Self' on a packed struct is always aligned; only field
  borrows are UB).
- question_mark (sensors.rs): rewrite the
  'if let Err(e) = ... { return Err(e); }' block as '... ?' while
  preserving the explanatory comment.

Verification:
  cargo clippy --lib         -> 0 warnings
  cargo clippy --all-targets -> 0 warnings
  cargo test --lib           -> 145/145 pass
  cargo build --release      -> ok

+++ app/rs/kr64/src/seccomp.rs
@@ -583,7 +583,7 @@ pub fn install() -> std::io::Result<()> {
         libc::syscall(
             libc::SYS_seccomp,
             SECCOMP_SET_MODE_FILTER as c_int,
-            SECCOMP_FILTER_FLAG_TSYNC as u32,
+            SECCOMP_FILTER_FLAG_TSYNC,
             &fprog as *const Fprog,
         )
     };
@@ -666,10 +666,10 @@ struct SigsysSiginfo {
 ///   2. Looks the syscall up in the trapped/killed sets.
 ///   3. If killed: log and call `_exit(1)` (no return).
 ///   4. If trapped: emulate the syscall by:
-///        a. Setting the return value register to 0 (success) — or
-///           `-ENOSYS` if we don't have a real emulation yet.
-///        b. Advancing the program counter past the syscall instruction
-///           so the kernel doesn't retry it.
+///      a. Setting the return value register to 0 (success) — or
+///      `-ENOSYS` if we don't have a real emulation yet.
+///      b. Advancing the program counter past the syscall instruction
+///      so the kernel doesn't retry it.
 ///   5. Returns; the kernel restores the (modified) context and
 ///      continues execution at the new PC.
 ///
